[refactor] 코드 리뷰 기반 코드 수정

+++ src/main/java/dev/woori/wooriBank/domain/transaction/transfer/service/TransferService.java
@@ -102,7 +102,7 @@ public class TransferService {
         historyRepository.save(
                 BankTransactionHistory.builder()
                         .account(account)
-                        .amount((int) amount)
+                        .amount(Math.toIntExact((long) amount))
                         .counterpartyName(counterparty)
                         .displayName(displayName)
                         .description(description)
